requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.30.1
pydantic==2.7.1
pydantic-settings==2.3.1
sentence-transformers==3.0.1
pinecone==5.0.1
python-multipart==0.0.9
pdfplumber==0.11.0
python-docx==1.1.2
numpy>=1.24.0,<2.0.0
httpx==0.27.0
pytest==8.2.0
pytest-asyncio==0.23.7
